Add Quickcheck testing (#51)
This PR is an example showing how we can use the `juvix.yaml` dependency
feature to write tests that depend on the current project and an
external dependency (in this case providing the quickcheck library).

Run `make test` from the root of the project. It will fetch the
quickcheck dependency, compile the Test module and run the tests from
the `test/` directory.