Add `Foldable` trait (#111)
- Closes #110.

# Polymorphic Foldable
This trait is equivalent to the `Foldable` class we all know in Haskell.
```
trait
type Foldable (f : Type -> Type) :=
  mkFoldable {-- Combine the elements of the type using the provided function starting with the element in the left-most position.
  foldl : {A B : Type} -> (B -> A -> B) -> B -> f A -> B};
```

# Monomorphic Foldable
This trait is similar to the
[`MonoFoldable`](https://hackage.haskell.org/package/mono-traversable-1.0.17.0/docs/Data-MonoTraversable.html#t:MonoFoldable)
class from the mono-traversable package. This trait is convenient
because it can be instantiated by monomorphic (or restricted) types such
as `Text`, `Bytes`, `Range N`.
```
trait
type Foldable (container elem : Type) :=
  mkFoldable {-- Combine the elements of the type using the provided function starting with the element in the left-most position.
  foldl : {B : Type} -> (B -> elem -> B) -> B -> container -> B};
```
This trait has two type parameters, `container` (e.g. `Text`) and `elem`
(e.g. `Char`).

All types that are instances of `PolymorphicFoldable`, should have
`Monomorphic` instances implemented via `fromPolymorphicFoldable`

As a consequence, the prelude will export the monomorphic version as
`Foldable` and the polymorphic version as `Polymorphic.Foldable`.


# Problems with having both Polymorphic and Monomorphic versions
## Limited types
Ideally, in the monomorphic version we'd want to compute `elem` from
`container` in some way. E.g.
```
trait
type FoldableImproved (container : Type) :=
  mkFoldable {
    elem : Type;
    foldl : {B : Type} -> (B -> elem -> B) -> B -> container -> B
  };
```
But this is not possible at the moment.

---------

Co-authored-by: Paul Cadman <git@paulcadman.dev>